+++ packages/stream_chat/lib/src/core/error/stream_chat_error.dart
@@ -1,5 +1,4 @@
 import 'package:equatable/equatable.dart';
-import 'package:stream_chat/src/core/error/chat_error_code.dart';
 import 'package:stream_chat/stream_chat.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';