+++ packages/stream_chat/CHANGELOG.md
@@ -9,6 +9,7 @@
 - [[#846]](https://github.com/GetStream/stream-chat-flutter/issues/846) Fixed `message.ownReactions` getting truncated when receiving a reaction event.
 - Add check for invalid image URLs
 - Fix `channelState.pinnedMessagesStream` getting reset to `0` after a channel update.
+- Fixed `unreadCount` after removing user from a channel.
 
 🔄 Changed
 
@@ -16,11 +17,11 @@
   new [edge server](https://getstream.io/blog/chat-edge-infrastructure) and will be removed in v4.0.0.
 - `channel.banUser`, `channel.unbanUser` is now deprecated in favor of the new `channel.banMember`
   and `channel.unbanMember`. These deprecated methods will be removed in v4.0.0.
-- Added banExpires property of type DateTime on the Member, OwnUser, and User models.
+- Added `banExpires` property of type `DateTime` on the `Member`, `OwnUser`, and `User` models.
+
 ✅ Added
 
 - Added `client.enrichUrl` endpoint for enriching URLs with metadata.
-- Fixed `unreadCount` after removing user from a channel.
 - Added `client.queryBannedUsers`, `channel.queryBannedUsers` endpoint for querying banned users.
 
 ## 3.3.1